Refactor module structure and add documentation
