+++ src/errors/mod.rs
@@ -1,2 +1,4 @@
+//! Contains data types for common errors encountered by the bot during runtime.
+
 mod bot;
 pub use bot::*;
